Update ex1-batch-transform.md high-level API
When I try to execute High-level Python Library code in a Sagemaker notebook with the conda_python3 kernel, I receive the error below. I suggested modifications that I think will work. I don't know what is intended in the transformer.transform() call. The ”/output” portion of that line doesn't seem like correct Python syntax to me.

It would be helpful if the example showed how to use the MNIST data and kmeans model that was used previously in the tutorial in a batch transformation.

---------------------------------------------------------------------------
NameError                                 Traceback (most recent call last)
<ipython-input-2-7e710140d094> in <module>()
      8 
      9 # Initialize the transformer object
---> 10 transformer = Transformer(model_name=my_model_name,
     11                         transform_job_name=job_name,
     12                         output_path="s3://{}/{}/Batch-Transform/".format(bucket,prefix)

NameError: name 'Transformer' is not defined